refactor tableexploreview
